+++ DirectA/packA.py
@@ -1,0 +1,6 @@
+import sys
+import os
+print os.getcwd()
+sys.path.append("..\ ")
+
+from ..DirectB import *
