+++ DirectB/PackB.py
@@ -1,0 +1,5 @@
+
+
+def Hello():
+    print "Hello World!"
+    return None
